refactor: standardise files to lowercase
